+++ src/locale/tr.js
@@ -8,7 +8,8 @@ export default {
     clear: 'Temizle',
     dateFormat: 'dd.MM.yyyy',
     timeFormat: 'hh:mm aa',
-    firstDay: 1
+    firstDay: 1,
+    selected: 'seçilmiş'
 };